% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_Xdir}
\alias{check_Xdir}
\title{Internal: check for valid computational direction arguments}
\usage{
check_Xdir(Xdir, Xopts, normalize = FALSE)
}
\arguments{
\item{Xdir}{a character or numeric indicating the direction}

\item{Xopts}{a character indicating the names of the vertices}

\item{normalize}{a logical if true Xdir is normalized to have an Aitchison norm of 1}
}
\value{
a numeric vector
}
\description{
Internal: check for valid computational direction arguments
}
\author{
Lukas Dargel
}
\keyword{internal}
