% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_Xvar}
\alias{check_Xvar}
\title{Internal: check for valid name of Xvar}
\usage{
check_Xvar(
  Xvar,
  trSry,
  return_type = c("NAME_SIMPLEX", "NAME_COORD", "pos")[1]
)
}
\arguments{
\item{Xvar}{a character or numeric indicating the direction}

\item{trSry}{a character indicating the names of the vertices}
}
\value{
a single integer or character
}
\description{
Users should always specify Xvar as "NAME_SIMPLEX",
which means before log-ratio transformations.
}
\author{
Lukas Dargel
}
\keyword{internal}
