% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDaImpact-package.R
\docType{package}
\name{CoDaImpact-package}
\alias{CoDaImpact}
\alias{CoDaImpact-package}
\title{CoDaImpact: Interpreting CoDa Regression Models}
\description{
Provides methods for interpreting CoDa (Compositional Data) regression models along the lines of "Pairwise share ratio interpretations of compositional regression models" (Dargel and Thomas-Agnan 2024) \doi{10.1016/j.csda.2024.107945}. The new methods include variation scenarios, elasticities, elasticity differences and share ratio elasticities. These tools are independent of log-ratio transformations and allow an interpretation in the original space of shares. 'CoDaImpact' is designed to be used with the 'compositions' package and its ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeCe/CoDaImpact}
  \item \url{https://lukece.github.io/CoDaImpact/}
  \item Report bugs at \url{https://github.com/LukeCe/CoDaImpact/issues}
}

}
\author{
\strong{Maintainer}: Lukas Dargel \email{lukas.dargel@mailbox.org} (\href{https://orcid.org/0000-0002-4839-506X}{ORCID})

Authors:
\itemize{
  \item Christine Thomas-Agnan \email{christine.thomas@tse-fr.eu} (\href{https://orcid.org/0000-0002-6430-3110}{ORCID})
}

Other contributors:
\itemize{
  \item Rodrigue Nasr [contributor]
  \item Sijia Pan [contributor]
  \item Iban Rendo Barreiro [contributor]
  \item Shuyao Li [contributor]
}

}
\keyword{internal}
