% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lmCoDa.R
\name{coef.lmCoDa}
\alias{coef.lmCoDa}
\title{Predictions, fitted values, residuals, and coefficients in CoDa models}
\usage{
\method{coef}{lmCoDa}(object, space = NULL, split = FALSE, ...)
}
\arguments{
\item{object}{class "lmCoDa"}

\item{space}{a character indicating in which space the prediction should
be returned. Supported are the options \code{c("clr", "simplex")}.}

\item{split}{logical, if \code{TRUE} the coefficients are reported as a list instead
of a matrix, where list structure reflects the explanatory variables of the model}

\item{...}{not used}
}
\value{
a matrix
}
\description{
These functions work as in the usual lm object.
They additionally offer the possibility use the \code{space} argument
which transforms them into directly into clr space or in the simplex.
}
\author{
Lukas Dargel
}
