% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct}
\alias{pct}
\title{Format numbers to percentages
This code copied from stats:::format.perc(), to avoid notes about the ::: operator.}
\usage{
pct(probs, digits = 10)
}
\description{
Format numbers to percentages
This code copied from stats:::format.perc(), to avoid notes about the ::: operator.
}
\keyword{internal}
