% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDaImpact-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{alr}
\alias{alrInv}
\alias{clr}
\alias{clrInv}
\alias{ilr}
\alias{ilrInv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{compositions}{\code{\link[compositions]{alr}}, \code{\link[compositions:alr]{alrInv}}, \code{\link[compositions]{clr}}, \code{\link[compositions:clr]{clrInv}}, \code{\link[compositions]{ilr}}, \code{\link[compositions:ilr]{ilrInv}}}
}}

