% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lmCoDa.R
\name{residuals.lmCoDa}
\alias{residuals.lmCoDa}
\title{Predictions, fitted values, residuals, and coefficients in CoDa models}
\usage{
\method{residuals}{lmCoDa}(object, space = NULL, ...)
}
\arguments{
\item{object}{class "lmCoDa"}

\item{space}{a character indicating in which space the prediction should
be returned. Supported are the options \code{c("clr", "simplex")}.}

\item{...}{passed on to \code{\link[=predict.lm]{predict.lm()}}}
}
\value{
matrix or vector
}
\description{
These functions work as in the usual lm object.
They additionally offer the possibility use the \code{space} argument
which transforms them into directly into clr space or in the simplex.
}
\author{
Lukas Dargel
}
