% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationTable.R
\name{VariationTable}
\alias{VariationTable}
\title{Effects of infinitesimal changes in CoDa models}
\usage{
VariationTable(
  object,
  Xvar,
  Xdir,
  obs = 1,
  inc_size = 0.1,
  inc_rate = NULL,
  Ytotal = 1
)
}
\arguments{
\item{object}{an object of class lmCoDa}

\item{Xvar}{a character indicating the name of the explanatory variable that changes}

\item{Xdir}{either character or numeric, to indicate the direction in which Xvar should change
\itemize{
\item when character this should be one of the components of X, in which case the direction is the corresponding vertex of the simplex
\item when numeric this argument is coerced to a unit vector in the simplex
\item (when Xvar refers to a scalar variable this argument is ignored)
}}

\item{obs}{a numeric indicating the observation used for the scenario}

\item{inc_size}{a numeric indicating the distance between each point in the scenario of X}

\item{inc_rate}{a numeric that can be used as an parametrization of the step size}

\item{Ytotal}{a numeric indicating the total of Y}
}
\value{
data.frame
}
\description{
This function allows to evaluate how a change in an explanatory variables
impacts the response variable in a CoDa regression model.
The changes are calculated based from the approximate formal presented
in Dargel and Thomas-Agnan (2023).
Changes in the response variables are provided as data.frame and the
underlying changes in the explanatory variable are given as attributes.
}
\details{
Developed in Dargel and Thomas-Agnan (2023)
}
\examples{

# XY-compositional model
res <- lmCoDa(
  ilr(cbind(left, right, extreme_right)) ~
  ilr(cbind(Educ_BeforeHighschool, Educ_Highschool, Educ_Higher)),
  data =  head(election, 20))

# Variation of age the education composition towards a summit ...
# (higher share of people with lower education)
VariationTable(res, Xvar = "cbind(Educ_BeforeHighschool, Educ_Highschool, Educ_Higher)", Xdir = "Educ_BeforeHighschool")

# The same changes using a compositional vector as direction
VariationTable(res, Xvar = "cbind(Educ_BeforeHighschool, Educ_Highschool, Educ_Higher)", Xdir = c(.5,.25,.25))

# Changes in a more general direction and for a diffrent observation
VariationTable(res, Xvar = "cbind(Educ_BeforeHighschool, Educ_Highschool, Educ_Higher)", Xdir = c(.35,.45,.10), obs = 2)
}
\author{
\itemize{
\item Lukas Dargel
\item Rodrigue Nasr
}
}
