% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDaImpact-package.R
\docType{package}
\name{CoDaImpact-package}
\alias{CoDaImpact}
\alias{CoDaImpact-package}
\title{CoDaImpact: Tools for interpreting compositional regression}
\description{
The package implements elasticities for interpreting CoDa regression models in the simplex. It is designed as an add-on to the composition package and its ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeCe/CoDaImpact}
  \item Report bugs at \url{https://github.com/LukeCe/CoDaImpact/issues}
}

}
\author{
\strong{Maintainer}: Lukas Dargel \email{lukas.dargel@mailbox.org} (\href{https://orcid.org/0000-0002-4839-506X}{ORCID})

Authors:
\itemize{
  \item Christine Thomas \email{christine.thomas@tse-fr.eu}
}

Other contributors:
\itemize{
  \item Rodrigue Nasr [contributor]
  \item Sijia Pan [contributor]
  \item Iban Rendo Barreiro [contributor]
  \item Shuyao Li [contributor]
}

}
\keyword{internal}
