% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDaImpact-package.R
\docType{package}
\name{CoDaImpact-package}
\alias{CoDaImpact}
\alias{CoDaImpact-package}
\title{CoDaImpact: Tools for interpreting CoDa regression models}
\description{
The CoDaImpact package implements additional tools for interpreting the results of CoDa regression models and is designed as an extension of the compositions package and its ecosystem. Our additional interpretation tools comprise variation scenarios, elasticities, elasticity differences and pairwise share-ratios increments. These tools have the advantage of being independent of log-ratio transformations and allow an interpretation in the original space of shares (for more details see Dargel and Thomas-Agnan (2023) \url{https://www.tse-fr.eu/publications/share-ratio-interpretations-compositional-regression-models}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeCe/CoDaImpact}
  \item \url{https://lukece.github.io/CoDaImpact/}
  \item Report bugs at \url{https://github.com/LukeCe/CoDaImpact/issues}
}

}
\author{
\strong{Maintainer}: Lukas Dargel \email{lukas.dargel@mailbox.org} (\href{https://orcid.org/0000-0002-4839-506X}{ORCID})

Authors:
\itemize{
  \item Christine Thomas \email{christine.thomas@tse-fr.eu}
}

Other contributors:
\itemize{
  \item Rodrigue Nasr [contributor]
  \item Sijia Pan [contributor]
  \item Iban Rendo Barreiro [contributor]
  \item Shuyao Li [contributor]
}

}
\keyword{internal}
